third_party/WebKit/LayoutTests/platform/android/http/tests/media/media-source/mediasource-config-change-mp4-v-bitrate-expected.txt

CONSOLE WARNING: line 106: Ignore this warning. See https://crbug.com/568704#c2
This is a testharness.js-based test.
PASS Tests mp4 video-only bitrate changes. 
Harness: the test ran to completion.
